do not load the level yaml file if it does not exist

+++ config/initializers/set_levels.rb
@@ -1,4 +1,6 @@
 module Eijiro
   level_file = File.join(Rails.root, %w(db level.yml))
-  LEVEL = YAML::load(File.open(level_file))
+  if File.exist?(level_file)
+    LEVEL = YAML::load(File.open(level_file))
+  end
 end
